+++ changelog.md
@@ -6,6 +6,8 @@
   ([#273](https://github.com/feltcoop/gro/pull/273))
 - fix writing source meta concurrently
   ([#288](https://github.com/feltcoop/gro/pull/288))
+- add `utils/throttleAsync.ts`
+  ([#288](https://github.com/feltcoop/gro/pull/288))
 
 ## 0.47.0